Refactor WPDB class and add nullability to Execute
The WPDB class has been refactored to remove the singleton pattern and instead each method now directly instantiates a new object. This change simplifies the logic in the WPDB class and leads to a more functional style of programming. In addition, the constructor in the Execute class is enhanced to accept null and handle the db connection appropriately.