Fix the handling of database error in GetResults.php
An error in the 'GetResults.php' file was resolved where the database error check was incorrectly calling '$wpdb->last_error' instead of using the object reference '$this->wpdb->last_error'. The change ensures that in case of a database error, a DatabaseException with the correct error message will be thrown.